Bugfix in mac os std::pow
git-svn-id: https://svn.liris.cnrs.fr/dgtal/trunk@953 5422497e-619f-4381-bc00-4cdae604c9f6

+++ tests/topology/testCellularGridSpaceND.cpp
@@ -115,11 +115,11 @@ bool testCellularGridSpaceND()
 
   trace.beginBlock ( "Testing faces in KSpace..." );
   Cells Nf = K.uFaces( center );
-  nbok += Nf.size() == round( pow( 3, K.dimension ) - 1 ) ? 1 : 0; 
+  nbok += Nf.size() == round( std::pow( 3.0 ,(int) K.dimension ) - 1 ) ? 1 : 0; 
   nb++;
   trace.info() << "(" << nbok << "/" << nb << ") "
 	       << Nf.size() << "(faces size) == "
-	       << round( pow( 3, K.dimension ) - 1 ) << "(3^dim()-1)" << endl;
+	       << round( std::pow( 3.0, (int)K.dimension ) - 1 ) << "(3^dim()-1)" << endl;
   trace.endBlock();
   
   trace.beginBlock ( "Testing block Incidence in KSpace..." );
